backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Basic Image Processing (install these first)
Pillow==10.2.0
numpy>=1.19.0,<1.25.0
opencv-python>=4.5.0

# Face Recognition (install dlib-binary first with: pip install dlib-binary)
# Then uncomment and install: pip install face-recognition
# face-recognition==1.3.0
